Added a Asynchonous implementation to the Socket Server.
So far many clients can connect to the server from any part of the network and send data to the sever independantly.

Modified the Client Application to ask the user  for the name of the PC which contains the server.

I'm still not a 100% sure how the asynchonous implementation works, more study is needed.